refactor tool list
